politics_pole/requirements.txt
requests
lxml
sqlalchemy
awesome-slugify
djangorestframework